harbor_export/tasks/mkt_s1_product_launch/environment/pyproject.toml
[project]
name = "long-horizon-gym"
version = "0.1.0"
description = "Long-Horizon Agent Evaluation Gym — scenario packages and shared core."
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.1",
    "openpyxl>=3.1.2",
    "reportlab>=4.0.0",
    "pypdf>=4.0.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40.0"]
openai = ["openai>=1.54.0"]
google = ["google-genai>=0.3.0"]

[tool.uv]
package = false

[tool.setuptools.packages.find]
include = ["gym_core*"]
